Change profile text length filter columns from BIGINT to SMALLINT

+++ migrations/2023-08-03-190519_initial_schema/up.sql
@@ -450,11 +450,10 @@ CREATE TABLE IF NOT EXISTS profile_state(
     profile_created_time_filter BIGINT,
     -- Filter setting for profile edited time in seconds.
     profile_edited_time_filter BIGINT,
-    -- TODO: Change to SMALLINT
     -- Filter setting for profile text min character count.
-    profile_text_min_characters_filter BIGINT,
+    profile_text_min_characters_filter SMALLINT,
     -- Filter setting for profile text max character count.
-    profile_text_max_characters_filter BIGINT,
+    profile_text_max_characters_filter SMALLINT,
     -- Profile iterator setting for random profile order.
     random_profile_order       BOOLEAN              NOT NULL    DEFAULT FALSE,
     latitude                   DOUBLE PRECISION               NOT NULL    DEFAULT 0.0,
